Get pkg/term to build for Solaris
