+++ pkg/term/tc_other.go
@@ -1,5 +1,6 @@
 // +build !windows
 // +build !linux !cgo
+// +build !solaris !cgo
 
 package term
