+++ pkg/term/term.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"os/signal"
 	"syscall"
-	"unsafe"
 )
 
 var (
@@ -45,27 +44,6 @@ func GetFdInfo(in interface{}) (uintptr, bool) {
 		isTerminalIn = IsTerminal(inFd)
 	}
 	return inFd, isTerminalIn
-}
-
-// GetWinsize returns the window size based on the specified file descriptor.
-func GetWinsize(fd uintptr) (*Winsize, error) {
-	ws := &Winsize{}
-	_, _, err := syscall.Syscall(syscall.SYS_IOCTL, fd, uintptr(syscall.TIOCGWINSZ), uintptr(unsafe.Pointer(ws)))
-	// Skip errno = 0
-	if err == 0 {
-		return ws, nil
-	}
-	return ws, err
-}
-
-// SetWinsize tries to set the specified window size for the specified file descriptor.
-func SetWinsize(fd uintptr, ws *Winsize) error {
-	_, _, err := syscall.Syscall(syscall.SYS_IOCTL, fd, uintptr(syscall.TIOCSWINSZ), uintptr(unsafe.Pointer(ws)))
-	// Skip errno = 0
-	if err == 0 {
-		return nil
-	}
-	return err
 }
 
 // IsTerminal returns true if the given file descriptor is a terminal.
